extract 'work start' endpoint logic to  library module as per standard pattern
